# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User Controller",
        "version": "0.1"
    },
    "paths": {
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of users",
                "operationId": "7e2f5799553ec93c9c43e2a58490d447",
                "responses": {
                    "200": {
                        "description": "List of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "operationId": "a1fce870c566eb5fc0cf3178ec5d2fed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cpf",
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "cpf": {
                                        "type": "integer",
                                        "example": "12345678910"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Itor Carlos"
                                    },
                                    "data_nascimento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2007-04-25"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Obtém informações de um usuário pelo CPF",
                "description": "Retorna os dados do usuário com base no CPF fornecido.",
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "cpf",
                        "in": "query",
                        "description": "CPF do usuário",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "12345678900"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "nome": {
                                            "type": "string",
                                            "example": "João Silva"
                                        },
                                        "cpf": {
                                            "type": "string",
                                            "example": "12345678900"
                                        },
                                        "data_nascimento": {
                                            "type": "string",
                                            "example": "1990-01-01"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "required": [
                    "cpf",
                    "name",
                    "data_nascimento"
                ],
                "properties": {
                    "cpf": {
                        "description": "CPF",
                        "type": "integer",
                        "example": 12345678910
                    },
                    "name": {
                        "description": "Name",
                        "type": "string",
                        "example": "Itor Carlos"
                    },
                    "data_nascimento": {
                        "description": "Data de nascimento",
                        "type": "string",
                        "format": "date",
                        "example": "2007-04-25"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Users",
            "description": "Users"
        },
        {
            "name": "Usuário",
            "description": "Usuário"
        }
    ]
}